Massive element interaction overhaul
Entirely changed the way that elements interact with  other elements. Now, Element.Step only performs movement, and after step, Element.Tick determines if an element should act on all of its neighbors or not. In order to make interactions less static, I need to now increase the complexity of the temperature simulation.